Removed no longer necessary modules
