+++ src/components/settlements/events/EventsContext.js
@@ -1,6 +1,6 @@
 import { createContext, useContext, useEffect, useState } from "react"
-import { getEvents, getSettlementEvents } from "../ApiManager"
-import { useSettlementForm } from "./SettlementFormContext"
+import { getEvents, getSettlementEvents } from "../../ApiManager"
+import { useSettlementForm } from "../form/SettlementFormContext"
 
 // Create the context variable using createContext
 const EventsContext = createContext()
